requirements.txt
streamlit
langchain
langchain-groq
python-dotenv
langchain_experimental
tabulate